Fixed, so that scripts would execute on remote machines

+++ modules/actions/kill_login_session.sh
@@ -18,12 +18,14 @@ if [[ "$#" -gt 1 ]]; then
     exit 1
 fi
 
+echo "Host: $(hostname -I)"
 SCRIPT_EXECUTOR_ID=$(id -u)
-#if [[ "$SCRIPT_EXECUTOR_ID" -ne 0 ]]; then
-#	echo "Please run $0 as root"
-#	echo "Exiting..."
-#	exit
-#fi
+echo "Running script as $(whoami) with ID: $SCRIPT_EXECUTOR_ID"
+if [[ "$SCRIPT_EXECUTOR_ID" -ne 0 ]]; then
+	echo "Please run $0 as root"
+	echo "Exiting..."
+	exit
+fi
 
 TTY_TO_KILL="$1"
 pkill -KILL -et $TTY_TO_KILL
